refactored structure, added re-exports
