+++ pyrio/exceptions/__init__.py
@@ -1,0 +1,5 @@
+from .exception import IllegalStateError as IllegalStateError
+from .exception import NullPointerError as NullPointerError
+from .exception import NoSuchElementError as NoSuchElementError
+from .exception import UnsupportedTypeError as UnsupportedTypeError
+from .exception import UnsupportedFileTypeError as UnsupportedFileTypeError
