+++ pyrio/streams/stream.py
@@ -1,7 +1,6 @@
-from pyrio.streams.base_stream import BaseStream
-from pyrio.iterators.generator import Generator
-from pyrio.iterators.itertools_mixin import ItertoolsMixin
-from pyrio.utils.decorator import pre_call, handle_consumed, dispatch
+from pyrio.decorators import pre_call, handle_consumed, dispatch
+from pyrio.streams import BaseStream
+from pyrio.iterators import StreamGenerator, ItertoolsMixin
 
 
 @pre_call(handle_consumed)
@@ -16,12 +15,12 @@ class Stream(BaseStream, ItertoolsMixin):
     @classmethod
     def iterate(cls, seed, operation, condition=None):
         """Creates infinite ordered Stream"""
-        return cls(Generator.iterate(seed, operation, condition))
+        return cls(StreamGenerator.iterate(seed, operation, condition))
 
     @classmethod
     def generate(cls, supplier):
         """Creates infinite unordered Stream with values generated by given supplier function"""
-        return cls(Generator.generate(supplier))
+        return cls(StreamGenerator.generate(supplier))
 
     @classmethod
     def constant(cls, element):
@@ -31,22 +30,20 @@ class Stream(BaseStream, ItertoolsMixin):
     @classmethod
     @dispatch(int, int, int)
     def from_range(cls, start, stop, step):
-        # TODO: fix docstr -> default step
         """Creates Stream from start (inclusive) to stop (exclusive) by an incremental step"""
-        return cls(Generator.range(start, stop, step))
+        return cls(StreamGenerator.range(start, stop, step))
 
     @classmethod
     @dispatch(int, int)
     def from_range(cls, start, stop):  # noqa: F811
-        # TODO: fix docstr -> default step
         """Creates Stream from start (inclusive) to stop (exclusive) by an incremental step of 1"""
-        return cls(Generator.range(start, stop, 1))
+        return cls(StreamGenerator.range(start, stop))
 
     @classmethod
     @dispatch(range)
     def from_range(cls, range_obj):  # noqa: F811
         """Creates Stream range object"""
-        return cls(Generator.range(range_obj.start, range_obj.stop, range_obj.step))
+        return cls(StreamGenerator.range(range_obj.start, range_obj.stop, range_obj.step))
 
     # NB: handle_consumed decorator needs access to toggle flag
     def take_nth(self, idx, default=None):
