+++ tests/test_optional.py
@@ -4,7 +4,7 @@ from contextlib import redirect_stdout
 import pytest
 
 from pyrio import Optional
-from pyrio.utils.exception import NoSuchElementError, NullPointerError
+from pyrio.exceptions import NoSuchElementError, NullPointerError
 
 
 def test_optional_get_raises():
